move dependency tracker
