+++ libs/server-sdk/src/data_components/dependency_tracker/data_kind.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <cstddef>
+
+namespace launchdarkly::server_side::data_components {
+enum class DataKind : std::size_t { kFlag = 0, kSegment = 1, kKindCount = 2 };
+}  // namespace launchdarkly::server_side::data_components
